+++ Magic/CHANGELOG.md
@@ -4,6 +4,7 @@
 
  - Magic mobs can now have a list of triggers, can cast spells or run commands on death/interval/damage.
  - Magic mobs will now focus on the player that has damaged them the most
+ - Non-trivial magic mobs will despawn on chunk unload and shutdown
  - Add support for clearing magic mobs from a specific location
  - Add a physical damage increase option to the wand buff shop for the Master Sword
  - Add physical protection option to buff shop